refactor(chat,langgraph): move chat-timeline primitive to chat, retarget to ChatAgentWithHistory
