+++ libs/chat/src/public-api.ts
@@ -41,6 +41,7 @@ export { ChatToolCallsComponent } from './lib/primitives/chat-tool-calls/chat-to
 export { ChatSubagentsComponent } from './lib/primitives/chat-subagents/chat-subagents.component';
 export { ChatThreadListComponent } from './lib/primitives/chat-thread-list/chat-thread-list.component';
 export type { Thread } from './lib/primitives/chat-thread-list/chat-thread-list.component';
+export { ChatTimelineComponent } from './lib/primitives/chat-timeline/chat-timeline.component';
 
 // DI provider
 export { provideChat, CHAT_CONFIG } from './lib/provide-chat';
